fix: history 없을 때 키워드 삭제 오류 해결
- history 없을 때 키워드 삭제가 안되는 오류를 해결함

+++ src/main/java/com/watchout/project/keyword/service/KeywordService.java
@@ -109,14 +109,16 @@ public class KeywordService {
         History history = historyRepositoryImpl.findHistoryByPhoneNumberAndKeyword(keywordDeleteRequestDto.getPhoneNumber(), keyword.getKeyword());
 
         User user = userRepositoryImpl.findUserByPhoneNumber(keywordDeleteRequestDto.getPhoneNumber());
-
         user.deleteKeyword(keyword);
 
-        history.deleteUser(user);
-        history.deleteKeyword();
+        if (history != null) {
+            history.deleteUser(user);
+            history.deleteKeyword();
+
+            historyRepository.delete(history);
+        }
 
         keywordRepository.delete(keyword);
-        historyRepository.delete(history);
 
         User updatedUser = userRepository.save(user);
